+++ src/material/menu/testing/BUILD.bazel
@@ -20,8 +20,8 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/overlay",
@@ -30,19 +30,6 @@ ng_test_library(
         "//src/cdk/testing/testbed",
         "//src/material/menu",
         "@npm//@angular/platform-browser",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/material/menu",
     ],
 )
